refactor(lib-sourcify): move legacy Vyper immutable IR helpers to their own module
Extract the IR-heuristic helpers added in this PR (returnLegacyVyperImmutableReferences,
isValidImmutableLength, and the internal IR walkers) out of VyperCompilation.ts into
legacyVyperImmutablesHelpers.ts. returnImmutableReferences stays in VyperCompilation and
imports the two it needs. Kept internal (not re-exported from index). No behavior change.
