+++ packages/lib-sourcify/src/Compilation/VyperCompilation.ts
@@ -11,7 +11,6 @@ import type {
   VyperJsonInput,
   VyperOutput,
   VyperOutputContract,
-  VyperIROutput,
   ImmutableReferences,
   LinkReferences,
 } from '@ethereum-sourcify/compilers-types';
@@ -22,6 +21,10 @@ import type {
   IVyperCompiler,
 } from './CompilationTypes';
 import { CompilationError } from './CompilationTypes';
+import {
+  isValidImmutableLength,
+  returnLegacyVyperImmutableReferences,
+} from './legacyVyperImmutablesHelpers';
 
 export function returnFixedVyperVersion(compilerVersion: string): string {
   if (semver.valid(compilerVersion)) {
@@ -74,107 +77,6 @@ export function returnAuxdataStyle(
     return AuxdataStyle.VYPER_LT_0_3_10;
   }
   return AuxdataStyle.VYPER;
-}
-
-const WORD_SIZE = 32;
-const LEGACY_VYPER_TEXT_IR_IMMUTABLE_RETURN =
-  /\[return,\s*(\d+),\s*\[add,\s*(\d+),\s*_lllsz\]\]/g;
-
-function isRecord(
-  value: VyperIROutput,
-): value is { [key: string]: VyperIROutput } {
-  return value !== null && typeof value === 'object' && !Array.isArray(value);
-}
-
-function isValidImmutableLength(length: VyperIROutput): length is number {
-  return (
-    typeof length === 'number' &&
-    Number.isSafeInteger(length) &&
-    length > 0 &&
-    length % WORD_SIZE === 0
-  );
-}
-
-function collectStructuredIrImmutableLengths(
-  node: VyperIROutput,
-  lengths: number[] = [],
-): number[] {
-  if (Array.isArray(node)) {
-    for (const child of node) {
-      collectStructuredIrImmutableLengths(child, lengths);
-    }
-    return lengths;
-  }
-
-  if (!isRecord(node)) {
-    return lengths;
-  }
-
-  const deployNode = node.deploy;
-  if (Array.isArray(deployNode) && isValidImmutableLength(deployNode[2])) {
-    lengths.push(deployNode[2]);
-  }
-
-  for (const child of Object.values(node)) {
-    collectStructuredIrImmutableLengths(child, lengths);
-  }
-  return lengths;
-}
-
-function extractTextIrImmutableLengths(ir: string): number[] {
-  const lengths: number[] = [];
-  for (const match of ir.matchAll(LEGACY_VYPER_TEXT_IR_IMMUTABLE_RETURN)) {
-    const length = Number(match[2]);
-    if (isValidImmutableLength(length)) {
-      lengths.push(length);
-    }
-  }
-  return lengths;
-}
-
-function getLegacyVyperImmutableLengthFromIr(
-  ir: VyperIROutput | undefined,
-): number | undefined {
-  if (ir === undefined) {
-    return undefined;
-  }
-  const lengths =
-    typeof ir === 'string'
-      ? extractTextIrImmutableLengths(ir)
-      : collectStructuredIrImmutableLengths(ir);
-
-  if (lengths.length !== 1) {
-    return undefined;
-  }
-  return lengths[0];
-}
-
-export function returnLegacyVyperImmutableReferences(
-  compilerOutput: VyperOutput | undefined,
-  compilationTarget: CompilationTarget,
-  runtimeBytecode: string,
-): ImmutableReferences {
-  const compilationTargetContract =
-    compilerOutput?.contracts?.[compilationTarget.path]?.[
-      compilationTarget.name
-    ];
-  const immutableLength = getLegacyVyperImmutableLengthFromIr(
-    compilationTargetContract?.ir,
-  );
-  if (immutableLength === undefined) {
-    return {};
-  }
-
-  const runtimeByteLength = runtimeBytecode.substring(2).length / 2;
-
-  return {
-    '0': [
-      {
-        length: immutableLength,
-        start: runtimeByteLength,
-      },
-    ],
-  };
 }
 
 export function returnImmutableReferences(
